Finalize Vector api
